% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_tree.R
\name{build_tree}
\alias{build_tree}
\title{Build dynamic tree using DiagrammeR and muir-generated data frame}
\usage{
build_tree(data, tree.dir = "LR", tree.height = NULL, tree.width = NULL)
}
\arguments{
\item{data}{A muir-generated data frame to be processed into a tree using DiagrammeR}

\item{tree.dir}{Direction of tree graph. Use either "LR" for left-to-right,
"RL" for right-to left, "TB" for top-to-bottom, or "BT" for bottom-to-top.}

\item{tree.height}{Control tree height to zoom in/out on nodes. Defaults to NULL}

\item{tree.width}{Control tree width to zoom in/out on nodes. Defaults to NULL}
}
\value{
An object of class \code{htmlwidget} (via DiagrammeR)that will
intelligently print itself into HTML in a variety of contexts
including the R console, within R Markdown documents,
and within Shiny output bindings.
}
\description{
This function allows users to easily and dynamically explore or document a
 dataset using a tree structure.
}

