% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s_dplyr.R
\name{eval.string.dplyr}
\alias{eval.string.dplyr}
\title{Internal function used by s_filter, s_select etc.}
\usage{
eval.string.dplyr(.data, .fun.name, ...)
}
\arguments{
\item{.data}{TBD}

\item{.fun.name}{TBD}

\item{...}{TBD}
}
\description{
Internal function used by s_filter, s_select etc.
}

