% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s_dplyr.R
\name{s_group_by}
\alias{s_group_by}
\title{Modified version of dplyr's group_by that uses string arguments}
\usage{
s_group_by(.data, ...)
}
\arguments{
\item{.data}{TBD}

\item{...}{TBD}
}
\description{
Modified version of dplyr's group_by that uses string arguments
}

