% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s_dplyr.R
\name{s_mutate}
\alias{s_mutate}
\title{Modified version of dplyr's arrange that uses string arguments}
\usage{
s_mutate(.data, ...)
}
\arguments{
\item{.data}{TBD}

\item{...}{TBD}
}
\description{
Modified version of dplyr's arrange that uses string arguments
}

