% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/muir.R
\name{muir}
\alias{muir}
\title{Explore Datasets with Trees}
\usage{
muir(data, node.levels, criteria, show.percent = TRUE, percent.digits = 2,
  show.empty.child = FALSE, ...)
}
\arguments{
\item{data}{A data frame to be explored}

\item{node.levels}{The columns from data that will be used to construct the tree in the order that
they should appear in the tree levels}

\item{criteria}{A data frame consisting of 4 character columns containing a column name (from node.levels),
an operator or function (e.g., "==",">", "is.na"), a value, and a label}

\item{show.percent}{Should nodes show the percent of records represented by that node compared to the total
number of records in data. Defaults to TRUE}

\item{percent.digits}{Remove and put digits in ...}

\item{show.empty.child}{Show a balanced tree with children nodes that are all empty or stop expanding
the tree once there is a parent node that empty. Defaults to FALSE -- don't show empty children nodes}

\item{...}{TBD OTHER PARAMS}
}
\value{
An object of class \code{htmlwidget} that will print itself as HTML and be saved as tree.html in wd()
}
\description{
This function allows users to easily and dynamically explore or document a dataset using a tree structure.
}

